refactor: 서비스 워커 등록 로직 리팩토링
